# refactor pickle support