--- a/api/response/pickle_test.go
+++ b/api/response/pickle_test.go
@@ -18,7 +18,7 @@ func BenchmarkHttpRespPickleEmptySeries(b *testing.B) {
 	}
 	var resp *Pickle
 	for n := 0; n < b.N; n++ {
-		resp = NewPickle(200, models.SeriesPickleFormat(data))
+		resp = NewPickle(200, models.SeriesByTarget(data))
 		resp.Body()
 		resp.Close()
 	}
@@ -34,7 +34,7 @@ func BenchmarkHttpRespPickleEmptySeriesNeedsEscaping(b *testing.B) {
 	}
 	var resp *Pickle
 	for n := 0; n < b.N; n++ {
-		resp = NewPickle(200, models.SeriesPickleFormat(data))
+		resp = NewPickle(200, models.SeriesByTarget(data))
 		resp.Body()
 		resp.Close()
 	}
@@ -58,7 +58,7 @@ func BenchmarkHttpRespPickleIntegers(b *testing.B) {
 	b.ResetTimer()
 	var resp *Pickle
 	for n := 0; n < b.N; n++ {
-		resp = NewPickle(200, models.SeriesPickleFormat(data))
+		resp = NewPickle(200, models.SeriesByTarget(data))
 		resp.Body()
 		resp.Close()
 	}
@@ -82,7 +82,7 @@ func BenchmarkHttpRespPickleFloats(b *testing.B) {
 	b.ResetTimer()
 	var resp *Pickle
 	for n := 0; n < b.N; n++ {
-		resp = NewPickle(200, models.SeriesPickleFormat(data))
+		resp = NewPickle(200, models.SeriesByTarget(data))
 		resp.Body()
 		resp.Close()
 	}
@@ -106,33 +106,8 @@ func BenchmarkHttpRespPickleNulls(b *testing.B) {
 	b.ResetTimer()
 	var resp *Pickle
 	for n := 0; n < b.N; n++ {
-		resp = NewPickle(200, models.SeriesPickleFormat(data))
+		resp = NewPickle(200, models.SeriesByTarget(data))
 		resp.Body()
 		resp.Close()
 	}
-}
-
-var foo []models.SeriesForPickle
-
-func BenchmarkConvertSeriesToPickleFormat10k(b *testing.B) {
-	points := make([]schema.Point, 10000)
-	baseTs := 1500000000
-	for i := 0; i < 10000; i++ {
-		points[i] = schema.Point{Val: 1.2345 * float64(i), Ts: uint32(baseTs + 10*i)}
-	}
-	data := []models.Series{
-		{
-			Target:     "foo",
-			Datapoints: points,
-			Interval:   10,
-		},
-	}
-	b.SetBytes(int64(len(points) * 12))
-
-	b.ResetTimer()
-	var bar []models.SeriesForPickle
-	for n := 0; n < b.N; n++ {
-		bar = models.SeriesPickleFormat(data)
-	}
-	foo = bar
 }