- use a "Picklable" interface that structs can implement.
- set datapoints to pickle.None type with the points are null (NaN)

fixes #615