@Dieterbe 2017-04-25
benchmarks look about the same (use grain of salt, benchmarks run on noisy cloud environment)

BEFORE (master)
```
BenchmarkHttpRespPickleEmptySeries-2                   	  300000	      4893 ns/op
BenchmarkHttpRespPickleEmptySeriesNeedsEscaping-2      	  300000	      4899 ns/op
BenchmarkHttpRespPickleIntegers-2                      	   10000	    115529 ns/op	 103.87 MB/s
BenchmarkHttpRespPickleFloats-2                        	   10000	    117172 ns/op	 102.41 MB/s
BenchmarkHttpRespPickleNulls-2                         	   10000	    115710 ns/op	 103.71 MB/s
BenchmarkConvertSeriesToPickleFormat10k-2              	   50000	     29210 ns/op	4108.18 MB/s
```

AFTER (this branch)
```
BenchmarkHttpRespPickleEmptySeries-2                   	  500000	      3713 ns/op
BenchmarkHttpRespPickleEmptySeriesNeedsEscaping-2      	  500000	      3727 ns/op
BenchmarkHttpRespPickleIntegers-2                      	   10000	    149439 ns/op	  80.30 MB/s
BenchmarkHttpRespPickleFloats-2                        	   10000	    149255 ns/op	  80.40 MB/s
BenchmarkHttpRespPickleNulls-2                         	   20000	     97434 ns/op	 123.16 MB/s
```